Categorization sheet form implementation (#7)
* Reorganise data structure

* Finish restructuring data, add fetching assignments with team input

* Simplify categorisation sheet - assignment relation

* Update README

* Introduce test data, reorganise data

* Add assignment types to sheet view

* Introduce app models, add parsing add enhance data modelling

* Introduce Dependency Injection

* Implement creating / updating categorization sheet form

* Group assignments on the categorization screen

* Introduce minimums for numeric assignments

* Introduce assignment groups' shares modal + points UI enhancements

* Introduce expected category calculations

+++ ScoutCamp/Code/Models/TeamCategorizationSheet.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import FirebaseFirestoreSwift
 
 struct TeamCategorizationSheet: FirebaseModel, Identifiable, Equatable, Hashable {
     let id: String
@@ -15,7 +14,7 @@ struct TeamCategorizationSheet: FirebaseModel, Identifiable, Equatable, Hashable
     let teamId: String
     let categoryId: String
 
-    let points: Int
+    let points: Double
     let isDraft: Bool
 
     let createdAt: Date
